split the utils files into file_utils and venv_utils for more clarity
